+++ app/controllers/doorkeeper/openid_connect/dynamic_client_registration_controller.rb
@@ -51,7 +51,7 @@ module Doorkeeper
         {
           name: params[:client_name],
           redirect_uri: params[:redirect_uris] || [],
-          scopes: params[:scope],
+          scopes: registration.permitted_scopes,
           confidential: registration.confidential_client?,
         }
       end